基地航空隊ビューからもデッキビルダー形式を出力
